+++ src/lib/plugins/autocomplete/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./autocomplete";
+export { DataProvider } from "./dataProvider";
